+++ .github/workflows/ci-cd.yaml
@@ -59,7 +59,7 @@ jobs:
       - name: Link Checker
         uses: lycheeverse/lychee-action@v1.9.0
         with:
-          fail: true
+          fail: false
 
   tests:
     name: Unit & Integration Tests
@@ -74,6 +74,51 @@ jobs:
           distribution: ${{ env.JDK_DISTRIBUTION }}
       - name: Run unit & integration tests
         run: mvn -B clean verify
+
+  docker-image:
+    name: Build Test & Release Development Docker Image
+    needs: tests
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+      - name: Set up JDK ${{ env.JDK_VERSION }}
+        uses: actions/setup-java@v3
+        with:
+          java-version: ${{ env.JDK_VERSION }}
+          distribution: ${{ env.JDK_DISTRIBUTION }}
+      - name: Build App WAR file so that Docker can pickup during image build
+        run: mvn clean package
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v2
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v2
+      - name: Test image build
+        uses: docker/build-push-action@v3
+        with:
+          context: .
+          push: false
+      - name: Login to DockerHub
+        if: github.ref == 'refs/heads/master'
+        uses: docker/login-action@v2
+        with:
+          username: jack20191124
+          password: ${{ secrets.DOCKERHUB_TOKEN }}
+      - name: Push image to DockerHub
+        if: github.ref == 'refs/heads/master'
+        uses: docker/build-push-action@v3
+        with:
+          context: .
+          push: true
+          tags: jack20191124/fiore:latest
+      - name: Push image Description
+        if: github.ref == 'refs/heads/master'
+        uses: peter-evans/dockerhub-description@v4
+        with:
+          username: jack20191124
+          password: ${{ secrets.DOCKERHUB_TOKEN }}
+          repository: jack20191124/fiore
+          readme-filepath: README.md
 
   documentation:
     name: Test & Release Documentation (including Javadoc)
@@ -112,49 +157,4 @@ jobs:
           enable_jekyll: false
           user_name: ${{ env.USER }}
           user_email: ${{ env.EMAIL }}
-          cname: chatbot.qubitpi.org
-
-  docker-image:
-    name: Build Test & Release Development Docker Image
-    needs: tests
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-      - name: Set up JDK ${{ env.JDK_VERSION }}
-        uses: actions/setup-java@v3
-        with:
-          java-version: ${{ env.JDK_VERSION }}
-          distribution: ${{ env.JDK_DISTRIBUTION }}
-      - name: Build App WAR file so that Docker can pickup during image build
-        run: mvn clean package
-      - name: Set up QEMU
-        uses: docker/setup-qemu-action@v2
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v2
-      - name: Test image build
-        uses: docker/build-push-action@v3
-        with:
-          context: .
-          push: false
-      - name: Login to DockerHub
-        if: github.ref == 'refs/heads/master'
-        uses: docker/login-action@v2
-        with:
-          username: jack20191124
-          password: ${{ secrets.DOCKERHUB_TOKEN }}
-      - name: Push image to DockerHub
-        if: github.ref == 'refs/heads/master'
-        uses: docker/build-push-action@v3
-        with:
-          context: .
-          push: true
-          tags: jack20191124/chatbot-ws:latest
-      - name: Push image Description
-        if: github.ref == 'refs/heads/master'
-        uses: peter-evans/dockerhub-description@v4
-        with:
-          username: jack20191124
-          password: ${{ secrets.DOCKERHUB_TOKEN }}
-          repository: jack20191124/chatbot-ws
-          readme-filepath: README.md
+          cname: fiore.qubitpi.org
