+++ Fiore-Project-intellij-code-style.xml
@@ -1,4 +1,4 @@
-<code_scheme name="chatbot-ws-Project">
+<code_scheme name="Fiore-Project">
   <option name="GENERATE_FINAL_LOCALS" value="true" />
   <option name="GENERATE_FINAL_PARAMETERS" value="true" />
   <option name="USE_EXTERNAL_ANNOTATIONS" value="true" />
