+++ README.md
@@ -1,14 +1,14 @@
-ChatbotWS
-=========
+Fiore
+=====
 
 ![Java Version Badge][Java Version Badge]
-[![GitHub Workflow Status][GitHub Workflow Status]](https://github.com/QubitPi/chatbot-ws/actions/workflows/ci-cd.yaml)
+[![GitHub Workflow Status][GitHub Workflow Status]](https://github.com/QubitPi/Fiore/actions/workflows/ci-cd.yaml)
 [![Apache License Badge]][Apache License, Version 2.0]
 
-__ChatbotWS__ is a [JSR 370] webservice template that lets us spin up Java webservice quickly through
-[GitHub templates]. Whether we want to create a chatbot that leverages popular APIs such as ChatGPT or connect to our
-own custom service, this template can do it all! Explore [chatbot.qubitpi.org](https://chatbot.qubitpi.org/) to view all
-of its available features, how to use them, examples and more!
+__Fiore__ is a [JSR 370] webservice template that lets us spin up Java webservice quickly through [GitHub templates].
+Whether we want to create a chatbot that leverages popular APIs such as ChatGPT or connect to our own custom service,
+this template can do it all! Explore [fiore.qubitpi.org](https://fiore.qubitpi.org/) to view all of its available features, how to use them,
+examples and more!
 
 ✨ Features
 -----------
@@ -37,7 +37,7 @@ commands:
 
 ```console
 export OPENAI_API_KEY=...
-docker run  -it -p 8080:8080 -e OPENAI_API_KEY=$OPENAI_API_KEY jack20191124/chatbot-ws
+docker run  -it -p 8080:8080 -e OPENAI_API_KEY=$OPENAI_API_KEY jack20191124/fiore
 ```
 
 That's it. A healthcheck endpoint can be pinned with
@@ -80,16 +80,15 @@ for details.
 License
 -------
 
-The use and distribution terms for [chatbot-ws]() are covered by the [Apache License, Version 2.0].
+The use and distribution terms for [fiore]() are covered by the [Apache License, Version 2.0].
 
 [Apache License Badge]: https://img.shields.io/badge/Apache%202.0-F25910.svg?style=for-the-badge&logo=Apache&logoColor=white
 [Apache License, Version 2.0]: https://www.apache.org/licenses/LICENSE-2.0
 
-[Documentation]: https://chatbot.qubitpi.org/
+[Documentation]: https://fiore.qubitpi.org/
 
 [GitHub templates]: https://docs.github.com/en/repositories/creating-and-managing-repositories/creating-a-template-repository#about-template-repositories
-[GitHub Workflow Status]: https://img.shields.io/github/actions/workflow/status/QubitPi/chatbot-ws/ci-cd.yaml?branch=master&logo=github&style=for-the-badge
+[GitHub Workflow Status]: https://img.shields.io/github/actions/workflow/status/QubitPi/Fiore/ci-cd.yaml?branch=master&logo=github&style=for-the-badge
 
 [Java Version Badge]: https://img.shields.io/badge/Java-17-brightgreen?style=for-the-badge&logo=OpenJDK&logoColor=white
-[chatbot-ws]: https://chatbot.qubitpi.org/
 [JSR 370]: https://jcp.org/en/jsr/detail?id=370
