+++ docs/README.md
@@ -1,5 +1,5 @@
-ChatbotWS Documentation
-====================
+Fiore Documentation
+===================
 
 This website is built using [Docusaurus 3](https://docusaurus.io/), a modern static website generator.
