+++ docs/blog/authors.yml
@@ -1,5 +1,5 @@
 jiaqi:
   name: Jiaqi Liu
-  title: Maintainer of ChatbotWS
+  title: Maintainer of Fiore
   url: https://github.com/QubitPi
   image_url: https://avatars.githubusercontent.com/u/16126939?v=4
