+++ pom.xml
@@ -5,13 +5,13 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.qubitpi</groupId>
-    <artifactId>chatbot-ws</artifactId>
+    <artifactId>fiore</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>war</packaging>
 
-    <name>ChatbotWS</name>
+    <name>Fiore</name>
     <description>Template project for building chatbot webservice running in Jetty</description>
-    <url>https://github.com/QubitPi/chatbot-ws</url>
+    <url>https://github.com/QubitPi/Fiore</url>
 
     <developers>
         <developer>
