+++ src/main/java/com/qubitpi/fiore/web/filters/CorsFilter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.qubitpi.chatbotws.web.filters;
+package com.qubitpi.fiore.web.filters;
 
 import jakarta.annotation.Priority;
 import jakarta.validation.constraints.NotNull;
