+++ src/test/groovy/com/qubitpi/fiore/application/ResourceConfigSpec.groovy
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.qubitpi.chatbotws.application
+package com.qubitpi.fiore.application
 
-import com.qubitpi.chatbotws.web.filters.CorsFilter
+import com.qubitpi.fiore.web.filters.CorsFilter
 
 import org.glassfish.jersey.internal.inject.Binder
