+++ src/test/groovy/com/qubitpi/fiore/web/filters/CorsFilterSpec.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.qubitpi.chatbotws.web.filters
+package com.qubitpi.fiore.web.filters
 
 import jakarta.ws.rs.container.ContainerRequestContext
 import jakarta.ws.rs.container.ContainerResponseContext
